… larq_compute_engine/tflite/tests/bconv2d_test.cc