## What do these changes do?

This PR adds optimised Indirect BGEMM kernels for Aarch64, which support float or int8 output. The kernels are written in a combination of Arm NEON intrinsics and inline assembly.

Concretely, three new kernels are added (ordered least to most efficient):
* `8x4x1_aarch64` - an 8x4 kernel that operates on depths in blocks of 1 (multiples of 32 input channels).
* `8x4x2_aarch64` - an 8x4 kernel that operates on depths in blocks of 2 (multiples of 64 input channels).
* `8x4x4_aarch64` - an 8x4 kernel that operates on depths in blocks of 4 (multiples of 128 input channels).

Additionally, each kernel has two variants: one for where the depth is exactly 1/2/4 (exactly 32/64/128 input channels), and one for any depth. The former case allows a slight optimisation as there is no need to loop over the depth dimension.

The kernel that's used is selected at runtime.

## How Has This Been Tested?

The existing kernel tests have been extended to cover more combinations of input/output channels. The 'big' kernel tests that don't run on CI pass locally for Aarch64 and Arm32.

## Benchmark Results

These benchmarks were taken with the LCE benchmarking tool on a Raspberry Pi 4B (running 64-bit Ubuntu) and on my Android phone (a OnePlus 6). To benchmark float outputs, I used the QuickNet models. To benchmark int8 outputs, I used the same QuickNet models but modified them to use Int8 tensors throughout.

All inference was run single-threaded as the Indirect BGEMM doesn't support multithreading yet. I used `num_runs=250` and report the average latency in ms and the standard deviation, for the im2col + BGEMM baseline and the kernels in this PR.

### Raspberry Pi 4B

| Model              | Baseline (im2col + BGEMM) | PR (indirect BGEMM) | % change |
|--------------------|:-------------------------:|:-------------------:|:--------:|
| QuickNet           | 30.04 +- 0.07             | 28.85 +- 0.06       | -4.0%    |
| QuickNetLarge      | 46.09 +- 0.08             | 43.74 +- 0.05       | -5.1%    |
| QuickNetXL         | 78.96 +- 0.09             | 75.22 +- 0.09       | -4.7%    |
| QuickNet Int8      | 33.96 +- 0.05             | 33.60 +- 0.05       | -1.1%    |
| QuickNetLarge Int8 | 48.75 +- 0.06             | 47.68 +- 0.04       | -2.2%    |
| QuickNetXL Int8    | 83.75 +- 0.06             | 82.07 +- 0.06       | -2.0%    |

### OnePlus 6

| Model              | Baseline (im2col + BGEMM) | PR (indirect BGEMM) | % change |
|--------------------|:-------------------------:|:-------------------:|:--------:|
| QuickNet           | 14.63 +- 0.06             | 14.31 +- 0.06       | -2.2%    |
| QuickNetLarge      | 22.07 +- 0.07             | 21.29 +- 0.06       | -3.5%    |
| QuickNetXL         | 38.14 +- 0.08             | 36.95 +- 0.08       | -3.1%    |
| QuickNet Int8      | 16.11 +- 0.07             | 15.89 +- 0.05       | -1.4%    |
| QuickNetLarge Int8 | 23.04 +- 0.06             | 22.47 +- 0.06       | -2.5%    |
| QuickNetXL Int8    | 40.05 +- 0.05             | 39.29 +- 0.06       | -1.9%    |

These results show a small but consistent improvement in latency, that's slightly more pronounced on the Raspberry Pi than the OnePlus 6, and is stronger for float output than int8 output.

## Related issue number

N/A.